api/views.py
from rest_framework import viewsets
from rest_framework.response import Response
from rest_framework import status
from .models import Noticia
from .serializers import NoticiaSerializer


class NoticiaViewSet(viewsets.ModelViewSet):
    queryset = Noticia.objects.all()
    serializer_class = NoticiaSerializer

    def list(self, request):
        queryset = self.get_queryset()
        serializer = self.get_serializer(queryset, many=True)
        return Response(serializer.data)

    def retrieve(self, request, pk=None):
        noticia = self.get_object()
        serializer = self.get_serializer(noticia)
        return Response(serializer.data)

    def create(self, request):
        serializer = self.get_serializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        self.perform_create(serializer)
        headers = self.get_success_headers(serializer.data)
        return Response(
            serializer.data,
            status=status.HTTP_201_CREATED,
            headers=headers)

    def update(self, request, pk=None):
        noticia = self.get_object()
        serializer = self.get_serializer(
            noticia, data=request.data, partial=True)
        serializer.is_valid(raise_exception=True)
        self.perform_update(serializer)
        return Response(serializer.data)

    def destroy(self, request, pk=None):
        noticia = self.get_object()
        self.perform_destroy(noticia)
        return Response(status=status.HTTP_204_NO_CONTENT)
