requirements.txt
# Django e Django REST Framework
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
flake8==7.1.1